Fixed the bug in GenericMigrationWrapperTest introduced with the latest changes in ServiceConfig.
GenericMigrationWrapperTest now reads its environment settings from a configuration file processed by ServiceConfig.

Fixed  a bug in DBMigrationPathFactory that caused the values of the preset setting parameters not to be read.

Fixed a bug in the GenericWrapperConfigFileExample.xml the command line in the example on how to use intermediate temporary files was broken, as the 'tr' command was supposed to read from a file - it is only capable of reading from standard input.